adds shader_mappings
