+++ include/Lib/opengl/shader.hpp
@@ -4,6 +4,7 @@
 
 #include "core.hpp"
 #include "utils.hpp"
+#include "shader_mappings.hpp"
 
 namespace GL
 {
@@ -35,6 +36,11 @@ namespace GL
 
 	struct ShaderProgram : OpenGLObject
 	{
+		vector<AttributeMapping> attribute_mappings;
+		vector<UniformMapping> uniform_mappings;
+		vector<UniformBlockMapping> uniform_block_mappings;
+		vector<StorageBlockMapping> storage_block_mappings;
+
 		CTOR(ShaderProgram, default)
 		COPY(ShaderProgram, delete)
 		MOVE(ShaderProgram, default)
@@ -61,54 +67,5 @@ namespace GL
 			for (auto & shader_stage: description.shader_stages)
 				glDetachShader(id, shader_stage->id);
 		}
-
-		struct AttributeMapping
-		{
-			u32 location;
-			GLenum glsl_type; // just for debug purposes
-			bool per_patch;
-			Geometry::Attribute::Key key;
-		};
-		vector<AttributeMapping> attribute_mappings;
-
-		struct UniformMapping
-		{
-			u32 location;
-			GLenum glsl_type; // just for debug purposes
-			std::string key;
-		};
-		vector<UniformMapping> uniform_mappings;
-
-		struct UniformBlockMapping
-		{
-			u32 location;
-			u32 data_size;
-			std::string key;
-
-			struct Variable
-			{
-				u32 offset;
-				GLenum glsl_type; // just for debug purposes
-				std::string key;
-			};
-			vector<Variable> variables;
-		};
-		vector<UniformBlockMapping> uniform_block_mappings;
-
-		struct StorageBlockMapping
-		{
-			u32 location;
-			u32 data_size;
-			std::string key;
-
-			struct Variable
-			{
-				u32 offset;
-				GLenum glsl_type; // just for debug purposes
-				std::string key;
-			};
-			vector<Variable> variables;
-		};
-		vector<StorageBlockMapping> storage_block_mappings;
 	};
 }
