+++ include/Lib/opengl/shader_mappings.hpp
@@ -1,0 +1,55 @@
+#pragma once
+
+#include "core.hpp"
+
+// TODO(bekorn): this dependency should be reversed, opengl should be contained in itself,
+//  geometry should build on top of opengl (or any other API later on)
+#include "Lib/geometry/core.hpp"
+
+namespace GL
+{
+	struct AttributeMapping
+	{
+		u32 location;
+		GLenum glsl_type; // just for debug purposes
+		bool per_patch;
+		Geometry::Attribute::Key key;
+	};
+
+	struct UniformMapping
+	{
+		u32 location;
+		GLenum glsl_type; // just for debug purposes
+		std::string key;
+	};
+
+	struct UniformBlockMapping
+	{
+		u32 location;
+		u32 data_size;
+		std::string key;
+
+		struct Variable
+		{
+			u32 offset;
+			GLenum glsl_type; // just for debug purposes
+			std::string key;
+		};
+		vector<Variable> variables;
+	};
+
+	struct StorageBlockMapping
+	{
+		u32 location;
+		u32 data_size;
+		std::string key;
+
+		struct Variable
+		{
+			u32 offset;
+			GLenum glsl_type; // just for debug purposes
+			std::string key;
+		};
+		vector<Variable> variables;
+	};
+}
