+++ source/main.cpp
@@ -1,19 +1,12 @@
 #include "Lib/core/.hpp"
 #include "Lib/opengl/.hpp"
+#include "Lib/opengl/use_dedicated_device_by_default.hpp"
 #include "Lib/glfw/.hpp"
 #include "Lib/imgui/.hpp"
 #include "Lib/asset_kitchen/assets.hpp"
 
 #include "game.hpp"
 #include "editor.hpp"
-
-// TODO(bekorn): move this back into opengl folder
-// To run the application with the external GPU, https://stackoverflow.com/a/39047129/2073225
-extern "C"
-{
-__declspec(dllexport) i64 NvOptimusEnablement = 1;
-__declspec(dllexport) i32 AmdPowerXpressRequestHighPerformance = 1;
-}
 
 i32 main(i32 argc, char** argv)
 {
